requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
djangorestframework==3.14.0
idna==3.6
MD5-Checksum-verifier==0.1.3
md5hash==0.0.6
Naked==0.1.32
pycryptodome==3.19.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
shellescape==3.8.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
